Tests/LibWeb/Text/expected/canvas/export.txt

1. "image/png"
2. "image/jpeg"
